feat :
Adds API key file and updates OCR page with secret API key

This commit adds a new file, `lib/data/api_key.dart`, which contains the API key. It also updates the `lib/views/pages/test_ocr2_page.dart` file to import the API key from the new file instead of hardcoding it. This ensures the API key is kept secret and not exposed in the code. This change improves security and allows for easier management of the API key.